+++ sharding-jdbc-orchestration/src/main/java/io/shardingsphere/jdbc/orchestration/internal/config/ConfigurationService.java
@@ -20,6 +20,7 @@ package io.shardingsphere.jdbc.orchestration.internal.config;
 import com.google.common.base.Strings;
 import io.shardingsphere.core.api.config.MasterSlaveRuleConfiguration;
 import io.shardingsphere.core.api.config.ShardingRuleConfiguration;
+import io.shardingsphere.core.exception.ShardingConfigurationException;
 import io.shardingsphere.core.rule.DataSourceParameter;
 import io.shardingsphere.jdbc.orchestration.internal.OrchestrationProxyConfiguration;
 import io.shardingsphere.jdbc.orchestration.internal.yaml.converter.DataSourceConverter;
@@ -30,6 +31,8 @@ import io.shardingsphere.jdbc.orchestration.internal.yaml.converter.ShardingConf
 import io.shardingsphere.jdbc.orchestration.reg.api.RegistryCenter;
 
 import javax.sql.DataSource;
+import java.util.HashMap;
+import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.Properties;
 
@@ -70,28 +73,28 @@ public final class ConfigurationService {
     
     private void persistDataSourceConfiguration(final Map<String, DataSource> dataSourceMap, final boolean isOverwrite) {
         if (isOverwrite || !hasDataSourceConfiguration()) {
+            if (dataSourceMap.isEmpty()) {
+                throw new ShardingConfigurationException("No available data source configuration for Orchestration.");
+            }
             regCenter.persist(configNode.getFullPath(ConfigurationNode.DATA_SOURCE_NODE_PATH), DataSourceConverter.dataSourceMapToYaml(dataSourceMap));
         }
     }
     
-    private void persistDataSourceParameterConfiguration(final Map<String, DataSourceParameter> dataSourceParameterMap, final boolean isOverwrite) {
-        if (isOverwrite || !hasDataSourceConfiguration()) {
-            regCenter.persist(configNode.getFullPath(ConfigurationNode.DATA_SOURCE_NODE_PATH), DataSourceParameterConverter.dataSourceParameterMapToYaml(dataSourceParameterMap));
-        }
-    }
-    
     private boolean hasDataSourceConfiguration() {
-        return regCenter.isExisted(configNode.getFullPath(ConfigurationNode.DATA_SOURCE_NODE_PATH));
+        return !Strings.isNullOrEmpty(regCenter.get(configNode.getFullPath(ConfigurationNode.DATA_SOURCE_NODE_PATH)));
     }
     
     private void persistShardingRuleConfiguration(final ShardingRuleConfiguration shardingRuleConfig, final boolean isOverwrite) {
         if (isOverwrite || !hasShardingRuleConfiguration()) {
+            if (shardingRuleConfig.getTableRuleConfigs().isEmpty()) {
+                throw new ShardingConfigurationException("No available sharding rule configuration for Orchestration.");
+            }
             regCenter.persist(configNode.getFullPath(ConfigurationNode.SHARDING_RULE_NODE_PATH), ShardingConfigurationConverter.shardingRuleConfigToYaml(shardingRuleConfig));
         }
     }
     
     private boolean hasShardingRuleConfiguration() {
-        return regCenter.isExisted(configNode.getFullPath(ConfigurationNode.SHARDING_RULE_NODE_PATH));
+        return !Strings.isNullOrEmpty(regCenter.get(configNode.getFullPath(ConfigurationNode.SHARDING_RULE_NODE_PATH)));
     }
     
     private void persistShardingConfigMap(final Map<String, Object> configMap, final boolean isOverwrite) {
@@ -101,7 +104,7 @@ public final class ConfigurationService {
     }
     
     private boolean hasShardingConfigMap() {
-        return regCenter.isExisted(configNode.getFullPath(ConfigurationNode.SHARDING_CONFIG_MAP_NODE_PATH));
+        return !Strings.isNullOrEmpty(regCenter.get(configNode.getFullPath(ConfigurationNode.SHARDING_CONFIG_MAP_NODE_PATH)));
     }
     
     private void persistShardingProperties(final Properties props, final boolean isOverwrite) {
@@ -111,7 +114,7 @@ public final class ConfigurationService {
     }
     
     private boolean hasShardingProperties() {
-        return regCenter.isExisted(configNode.getFullPath(ConfigurationNode.SHARDING_PROPS_NODE_PATH));
+        return !Strings.isNullOrEmpty(regCenter.get(configNode.getFullPath(ConfigurationNode.SHARDING_PROPS_NODE_PATH)));
     }
     
     /**
@@ -134,12 +137,15 @@ public final class ConfigurationService {
     
     private void persistMasterSlaveRuleConfiguration(final MasterSlaveRuleConfiguration masterSlaveRuleConfig, final boolean isOverwrite) {
         if (isOverwrite || !hasMasterSlaveRuleConfiguration()) {
+            if (masterSlaveRuleConfig.getMasterDataSourceName().isEmpty()) {
+                throw new ShardingConfigurationException("No available master slave configuration for Orchestration.");
+            }
             regCenter.persist(configNode.getFullPath(ConfigurationNode.MASTER_SLAVE_RULE_NODE_PATH), MasterSlaveConfigurationConverter.masterSlaveRuleConfigToYaml(masterSlaveRuleConfig));
         }
     }
     
     private boolean hasMasterSlaveRuleConfiguration() {
-        return regCenter.isExisted(configNode.getFullPath(ConfigurationNode.MASTER_SLAVE_RULE_NODE_PATH));
+        return !Strings.isNullOrEmpty(regCenter.get(configNode.getFullPath(ConfigurationNode.MASTER_SLAVE_RULE_NODE_PATH)));
     }
     
     private void persistMasterSlaveConfigMap(final Map<String, Object> configMap, final boolean isOverwrite) {
@@ -149,7 +155,7 @@ public final class ConfigurationService {
     }
     
     private boolean hasMasterSlaveConfigMap() {
-        return regCenter.isExisted(configNode.getFullPath(ConfigurationNode.MASTER_SLAVE_CONFIG_MAP_NODE_PATH));
+        return !Strings.isNullOrEmpty(regCenter.get(configNode.getFullPath(ConfigurationNode.MASTER_SLAVE_CONFIG_MAP_NODE_PATH)));
     }
     
     private void persistMasterSlaveProperties(final Properties props, final boolean isOverwrite) {
@@ -159,7 +165,7 @@ public final class ConfigurationService {
     }
     
     private boolean hasMasterSlaveProperties() {
-        return regCenter.isExisted(configNode.getFullPath(ConfigurationNode.MASTER_SLAVE_PROPS_NODE_PATH));
+        return !Strings.isNullOrEmpty(regCenter.get(configNode.getFullPath(ConfigurationNode.MASTER_SLAVE_PROPS_NODE_PATH)));
     }
     
     /**
@@ -173,12 +179,24 @@ public final class ConfigurationService {
         persistProxyRuleConfiguration(orchestrationProxyConfiguration, isOverwrite);
     }
     
+    private void persistDataSourceParameterConfiguration(final Map<String, DataSourceParameter> dataSourceParameterMap, final boolean isOverwrite) {
+        if (isOverwrite || !hasDataSourceConfiguration()) {
+            if (dataSourceParameterMap.isEmpty()) {
+                throw new ShardingConfigurationException("No available data source configuration for Orchestration.");
+            }
+            regCenter.persist(configNode.getFullPath(ConfigurationNode.DATA_SOURCE_NODE_PATH), DataSourceParameterConverter.dataSourceParameterMapToYaml(dataSourceParameterMap));
+        }
+    }
+    
     private boolean hasProxyConfig() {
-        return regCenter.isExisted(configNode.getFullPath(ConfigurationNode.PROXY_RULE_NODE_PATH));
+        return !Strings.isNullOrEmpty(regCenter.get(configNode.getFullPath(ConfigurationNode.PROXY_RULE_NODE_PATH)));
     }
     
     private void persistProxyRuleConfiguration(final OrchestrationProxyConfiguration orchestrationProxyConfiguration, final boolean isOverwrite) {
         if (isOverwrite || !hasProxyConfig()) {
+            if (orchestrationProxyConfiguration.getShardingRule().getTables().isEmpty() && orchestrationProxyConfiguration.getMasterSlaveRule().getMasterDataSourceName().isEmpty()) {
+                throw new ShardingConfigurationException("No available proxy rule configuration for Orchestration.");
+            }
             regCenter.persist(configNode.getFullPath(ConfigurationNode.PROXY_RULE_NODE_PATH), ProxyConfigurationConverter.proxyConfigToYaml(orchestrationProxyConfiguration));
         }
     }
@@ -189,7 +207,11 @@ public final class ConfigurationService {
      * @return data source configuration map
      */
     public Map<String, DataSource> loadDataSourceMap() {
-        return DataSourceConverter.dataSourceMapFromYaml(regCenter.getDirectly(configNode.getFullPath(ConfigurationNode.DATA_SOURCE_NODE_PATH)));
+        try {
+            return DataSourceConverter.dataSourceMapFromYaml(regCenter.getDirectly(configNode.getFullPath(ConfigurationNode.DATA_SOURCE_NODE_PATH)));
+        } catch (final Exception ex) {
+            throw new ShardingConfigurationException("No available data source configuration to load.");
+        }
     }
     
     /**
@@ -198,7 +220,11 @@ public final class ConfigurationService {
      * @return data sources map
      */
     public Map<String, DataSourceParameter> loadDataSources() {
-        return DataSourceParameterConverter.dataSourceParameterMapFromYaml(regCenter.getDirectly(configNode.getFullPath(ConfigurationNode.DATA_SOURCE_NODE_PATH)));
+        try {
+            return DataSourceParameterConverter.dataSourceParameterMapFromYaml(regCenter.getDirectly(configNode.getFullPath(ConfigurationNode.DATA_SOURCE_NODE_PATH)));
+        } catch (final Exception ex) {
+            throw new ShardingConfigurationException("No available data source configuration to load.");
+        }
     }
     
     /**
@@ -207,7 +233,11 @@ public final class ConfigurationService {
      * @return sharding rule configuration
      */
     public ShardingRuleConfiguration loadShardingRuleConfiguration() {
-        return ShardingConfigurationConverter.shardingRuleConfigFromYaml(regCenter.getDirectly(configNode.getFullPath(ConfigurationNode.SHARDING_RULE_NODE_PATH)));
+        try {
+            return ShardingConfigurationConverter.shardingRuleConfigFromYaml(regCenter.getDirectly(configNode.getFullPath(ConfigurationNode.SHARDING_RULE_NODE_PATH)));
+        } catch (final Exception ex) {
+            throw new ShardingConfigurationException("No available sharding rule configuration to load.");
+        }
     }
     
     /**
@@ -217,7 +247,8 @@ public final class ConfigurationService {
      */
     @SuppressWarnings("unchecked")
     public Map<String, Object> loadShardingConfigMap() {
-        return ShardingConfigurationConverter.configMapFromYaml(regCenter.getDirectly(configNode.getFullPath(ConfigurationNode.SHARDING_CONFIG_MAP_NODE_PATH)));
+        String data = regCenter.getDirectly(configNode.getFullPath(ConfigurationNode.SHARDING_CONFIG_MAP_NODE_PATH));
+        return Strings.isNullOrEmpty(data) ? new HashMap<String, Object>() : ShardingConfigurationConverter.configMapFromYaml(data);
     }
     
     /**
@@ -236,7 +267,11 @@ public final class ConfigurationService {
      * @return master-slave rule configuration
      */
     public MasterSlaveRuleConfiguration loadMasterSlaveRuleConfiguration() {
-        return MasterSlaveConfigurationConverter.masterSlaveRuleConfigFromYaml(regCenter.getDirectly(configNode.getFullPath(ConfigurationNode.MASTER_SLAVE_RULE_NODE_PATH)));
+        try {
+            return MasterSlaveConfigurationConverter.masterSlaveRuleConfigFromYaml(regCenter.getDirectly(configNode.getFullPath(ConfigurationNode.MASTER_SLAVE_RULE_NODE_PATH)));
+        } catch (final Exception ex) {
+            throw new ShardingConfigurationException("No available master slave rule configuration to load.");
+        }
     }
     
     /**
@@ -246,7 +281,8 @@ public final class ConfigurationService {
      */
     @SuppressWarnings("unchecked")
     public Map<String, Object> loadMasterSlaveConfigMap() {
-        return MasterSlaveConfigurationConverter.configMapFromYaml(regCenter.getDirectly(configNode.getFullPath(ConfigurationNode.MASTER_SLAVE_CONFIG_MAP_NODE_PATH)));
+        String data = regCenter.getDirectly(configNode.getFullPath(ConfigurationNode.MASTER_SLAVE_CONFIG_MAP_NODE_PATH));
+        return Strings.isNullOrEmpty(data) ? new LinkedHashMap<String, Object>() : MasterSlaveConfigurationConverter.configMapFromYaml(data);
     }
     
     /**
@@ -265,6 +301,10 @@ public final class ConfigurationService {
      * @return proxy configuration
      */
     public OrchestrationProxyConfiguration loadProxyConfiguration() {
-        return ProxyConfigurationConverter.proxyConfigFromYaml(regCenter.getDirectly(configNode.getFullPath(ConfigurationNode.PROXY_RULE_NODE_PATH)));
+        try {
+            return ProxyConfigurationConverter.proxyConfigFromYaml(regCenter.getDirectly(configNode.getFullPath(ConfigurationNode.PROXY_RULE_NODE_PATH)));
+        } catch (final Exception ex) {
+            throw new ShardingConfigurationException("No available proxy configuration to load.");
+        }
     }
 }